+++ src/main/java/com/study/battleq/infrastructure/config/jwt/CustomAuthenticationEntryPoint.java
@@ -1,4 +1,4 @@
-package com.study.battleq.infrastructure.jwt;
+package com.study.battleq.infrastructure.config.jwt;
 
 import org.springframework.security.core.AuthenticationException;
 import org.springframework.security.web.AuthenticationEntryPoint;
